requirements.txt
python-telegram-bot==20.7
APScheduler==3.10.4
python-dotenv==1.0.0
google-generativeai==0.8.3
requests==2.31.0
pytz==2023.3
pyttsx3==2.90
Pillow==10.1.0
email-validator==2.1.0 